+++ tests/testthat/test-predict.R
@@ -1,8 +1,8 @@
 # Gaussian ----
 
 test_that("eglm (gaussian) + cholesky singularity.method == glm", {
-  m1 <- glm(mpg ~ wt + am, family = gaussian(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = gaussian(), data = mtcars, singularity.method = "Cholesky")
+  m1 <- glm(mpg ~ wt + am, family = gaussian, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = gaussian, data = mtcars, singularity.method = "Cholesky")
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -17,8 +17,8 @@ test_that("eglm (gaussian) + cholesky singularity.method == glm", {
 # Inverse-Gaussian ----
 
 test_that("eglm (inverse.gaussian) == glm", {
-  m1 <- glm(mpg ~ wt + am, family = inverse.gaussian(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = inverse.gaussian(), data = mtcars)
+  m1 <- glm(mpg ~ wt + am, family = inverse.gaussian, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = inverse.gaussian, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -33,8 +33,8 @@ test_that("eglm (inverse.gaussian) == glm", {
 # Gamma ----
 
 test_that("eglm (gaussian) == glm", {
-  m1 <- glm(mpg ~ wt + am, family = Gamma(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = Gamma(), data = mtcars)
+  m1 <- glm(mpg ~ wt + am, family = Gamma, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = Gamma, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -49,8 +49,8 @@ test_that("eglm (gaussian) == glm", {
 # Binomial ----
 
 test_that("eglm (binomial) == glm", {
-  m1 <- glm(am ~ wt + mpg, family = binomial(), data = mtcars)
-  m2 <- eglm(am ~ wt + mpg, family = binomial(), data = mtcars)
+  m1 <- glm(am ~ wt + mpg, family = binomial, data = mtcars)
+  m2 <- eglm(am ~ wt + mpg, family = binomial, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -65,8 +65,8 @@ test_that("eglm (binomial) == glm", {
 # Quasi-Binomial ----
 
 test_that("eglm (quasibinomial) == glm", {
-  m1 <- glm(am ~ wt + mpg, family = quasibinomial(), data = mtcars)
-  m2 <- eglm(am ~ wt + mpg, family = quasibinomial(), data = mtcars)
+  m1 <- glm(am ~ wt + mpg, family = quasibinomial, data = mtcars)
+  m2 <- eglm(am ~ wt + mpg, family = quasibinomial, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -81,8 +81,8 @@ test_that("eglm (quasibinomial) == glm", {
 # Poisson ----
 
 test_that("eglm (poisson) == glm", {
-  m1 <- glm(am ~ wt + mpg, family = poisson(), data = mtcars)
-  m2 <- eglm(am ~ wt + mpg, family = poisson(), data = mtcars)
+  m1 <- glm(am ~ wt + mpg, family = poisson, data = mtcars)
+  m2 <- eglm(am ~ wt + mpg, family = poisson, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -97,8 +97,8 @@ test_that("eglm (poisson) == glm", {
 # Quasi-Poisson ----
 
 test_that("eglm (quasipoisson) == glm", {
-  m1 <- glm(mpg ~ wt + am, family = quasipoisson(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = quasipoisson(), data = mtcars)
+  m1 <- glm(mpg ~ wt + am, family = quasipoisson, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = quasipoisson, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -113,8 +113,8 @@ test_that("eglm (quasipoisson) == glm", {
 # Quasi ----
 
 test_that("eglm (quasi) == glm", {
-  m1 <- glm(mpg ~ wt + am, family = quasi(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = quasi(), data = mtcars)
+  m1 <- glm(mpg ~ wt + am, family = quasi, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = quasi, data = mtcars)
 
   expect_equal(
     predict(m1, newdata = mtcars, type = "link"),
@@ -129,8 +129,8 @@ test_that("eglm (quasi) == glm", {
 # NULL warnings ----
 
 test_that("predict.eglm fails with NULL fitted values + newdata", {
-  m1 <- eglm(mpg ~ wt + am, family = gaussian(), data = mtcars)
-  m2 <- eglm(mpg ~ wt + am, family = gaussian(), data = mtcars)
+  m1 <- eglm(mpg ~ wt + am, family = gaussian, data = mtcars)
+  m2 <- eglm(mpg ~ wt + am, family = gaussian, data = mtcars)
 
   expect_equal(predict(m1), predict(m2))
   m1$fitted.values <- NULL