+++ tests/testthat/test-summary-eglm.R
@@ -1,6 +1,6 @@
 test_that("gaussian summary is equivalent to glm", {
-  m1 <- summary(glm(mpg ~ wt, family = gaussian(), data = mtcars))
-  m2 <- summary(eglm(mpg ~ wt, family = gaussian(), data = mtcars))
+  m1 <- summary(glm(mpg ~ wt, family = gaussian, data = mtcars))
+  m2 <- summary(eglm(mpg ~ wt, family = gaussian, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- as.data.frame(m2$coefficients)
   em2$`Pr(>|t|)` <- as.numeric(em2$`Pr(>|t|)`)
@@ -26,8 +26,8 @@ test_that("gaussian summary is equivalent to glm", {
 })
 
 test_that("inverse.gaussian summary is equivalent to glm", {
-  m1 <- summary(glm(mpg ~ wt, family = inverse.gaussian(), data = mtcars))
-  m2 <- summary(eglm(mpg ~ wt, family = inverse.gaussian(), data = mtcars))
+  m1 <- summary(glm(mpg ~ wt, family = inverse.gaussian, data = mtcars))
+  m2 <- summary(eglm(mpg ~ wt, family = inverse.gaussian, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- as.data.frame(m2$coefficients)
   em2$`Pr(>|t|)` <- as.numeric(em2$`Pr(>|t|)`)
@@ -53,8 +53,8 @@ test_that("inverse.gaussian summary is equivalent to glm", {
 })
 
 test_that("binomial summary is equivalent to glm", {
-  m1 <- summary(glm(am ~ mpg, family = binomial(), data = mtcars))
-  m2 <- summary(eglm(am ~ mpg, family = binomial(), data = mtcars))
+  m1 <- summary(glm(am ~ mpg, family = binomial, data = mtcars))
+  m2 <- summary(eglm(am ~ mpg, family = binomial, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- round(as.data.frame(m2$coefficients), 4)
   expect_equal(m1$call$formula, m2$call$formula)
@@ -78,8 +78,8 @@ test_that("binomial summary is equivalent to glm", {
 })
 
 test_that("quasi-binomial summary is equivalent to glm", {
-  m1 <- summary(glm(am ~ mpg, family = quasibinomial(), data = mtcars))
-  m2 <- summary(eglm(am ~ mpg, family = quasibinomial(), data = mtcars))
+  m1 <- summary(glm(am ~ mpg, family = quasibinomial, data = mtcars))
+  m2 <- summary(eglm(am ~ mpg, family = quasibinomial, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- round(as.data.frame(m2$coefficients), 4)
   expect_equal(m1$call$formula, m2$call$formula)
@@ -103,8 +103,8 @@ test_that("quasi-binomial summary is equivalent to glm", {
 })
 
 test_that("poisson summary is equivalent to glm", {
-  m1 <- summary(glm(am ~ mpg, family = poisson(), data = mtcars))
-  m2 <- summary(eglm(am ~ mpg, family = poisson(), data = mtcars))
+  m1 <- summary(glm(am ~ mpg, family = poisson, data = mtcars))
+  m2 <- summary(eglm(am ~ mpg, family = poisson, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- round(as.data.frame(m2$coefficients), 4)
   expect_equal(m1$call$formula, m2$call$formula)
@@ -128,8 +128,8 @@ test_that("poisson summary is equivalent to glm", {
 })
 
 test_that("quasi-poisson summary is equivalent to glm", {
-  m1 <- summary(glm(am ~ mpg, family = quasipoisson(), data = mtcars))
-  m2 <- summary(eglm(am ~ mpg, family = quasipoisson(), data = mtcars))
+  m1 <- summary(glm(am ~ mpg, family = quasipoisson, data = mtcars))
+  m2 <- summary(eglm(am ~ mpg, family = quasipoisson, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- round(as.data.frame(m2$coefficients), 4)
   expect_equal(m1$call$formula, m2$call$formula)
@@ -153,8 +153,8 @@ test_that("quasi-poisson summary is equivalent to glm", {
 })
 
 test_that("quasi summary is equivalent to glm", {
-  m1 <- summary(glm(am ~ mpg, family = quasi(), data = mtcars))
-  m2 <- summary(eglm(am ~ mpg, family = quasi(), data = mtcars))
+  m1 <- summary(glm(am ~ mpg, family = quasi, data = mtcars))
+  m2 <- summary(eglm(am ~ mpg, family = quasi, data = mtcars))
   em1 <- round(as.data.frame(m1$coefficients), 4)
   em2 <- round(as.data.frame(m2$coefficients), 4)
   expect_equal(m1$call$formula, m2$call$formula)